fix: remove duplicate built logger file
This file was shadowing the real file, and getting imported instead.
I'm guessing this file got moved at some point, but the built file
did not get removed.

This fixes a few issues:
- The log.file option was being ignored before
- The max size of the log file was 1M instead of 50M
- The log was silent if NODE_ENV == "test"

These are all changes that were made in the real built file, but
were not being picked up.